Move env/config utility functions from glue to new utils module
